+++ src/index.ts
@@ -1,2 +1,7 @@
-export { StatewaveClient, StatewaveAPIError, StatewaveConnectionError } from "./client.js";
+export {
+  StatewaveClient,
+  StatewaveAPIError,
+  StatewaveConnectionError,
+  StatewaveUnreplayableError,
+} from "./client.js";
 export type * from "./types.js";